Refactor namespaces and update model references
Changed the namespace from `TourBooking.ApiService.Contracts` to `TourBooking.ApiService.Contracts.Models` across multiple files to better organize data models related to tours. Updated `ToursApiClient`, `ToursContext`, `CreateTour.razor`, and `TourDtoFactory` to reference the new models namespace. Modified the `Produces` method in `ToursEndpoints.cs` for more generalized response handling in the `GetAllTours` endpoint.
