+++ src/TourBooking.Web/Components/Pages/CreateTour.razor
@@ -2,6 +2,7 @@
 @rendermode InteractiveServer
 @using System.ComponentModel.DataAnnotations
 @using TourBooking.ApiService.Contracts
+@using TourBooking.ApiService.Contracts.Models
 @using TourBooking.Web.Contracts
 @inject ToursApiClient ToursApi
